+++ src/equations/laplace_diffusion.jl
@@ -30,6 +30,58 @@ see [`StepsizeCallback`](@ref).
     return equations_parabolic.diffusivity
 end
 
+@inline function penalty(u_outer, u_inner,
+                         equations_parabolic::AbstractLaplaceDiffusion,
+                         ::ParabolicFormulationLocalDG{Nothing})
+    return zero(u_inner)
+end
+
+@inline function penalty(u_outer, u_inner,
+                         equations_parabolic::AbstractLaplaceDiffusion,
+                         dg::ParabolicFormulationLocalDG)
+    return dg.penalty_parameter .* (u_outer - u_inner) .*
+           equations_parabolic.diffusivity
+end
+
+@inline function penalty(u_outer, u_inner, inv_h,
+                         equations_parabolic::AbstractLaplaceDiffusion,
+                         dg::ParabolicFormulationLocalDG)
+    return inv_h .* penalty(u_outer, u_inner, equations_parabolic, dg)
+end
+
+@inline function penalty_diffusivity(u_outer, u_inner, x, t,
+                                     equations_parabolic::AbstractLaplaceDiffusion)
+    return penalty_diffusivity(u_outer, u_inner, x, t,
+                               have_constant_diffusivity(equations_parabolic),
+                               equations_parabolic)
+end
+
+@inline function penalty_diffusivity(u_outer, u_inner, x, t, ::True,
+                                     equations_parabolic::AbstractLaplaceDiffusion)
+    return equations_parabolic.diffusivity
+end
+
+@inline function penalty_diffusivity(u_outer, u_inner, x, t, ::False,
+                                     equations_parabolic::AbstractLaplaceDiffusion)
+    diffusivity_outer = max_diffusivity(u_outer, x, t, equations_parabolic)
+    diffusivity_inner = max_diffusivity(u_inner, x, t, equations_parabolic)
+    return max.(diffusivity_outer, diffusivity_inner)
+end
+
+@inline function penalty(u_outer, u_inner, inv_h, x, t,
+                         equations_parabolic::AbstractLaplaceDiffusion,
+                         ::ParabolicFormulationLocalDG{Nothing})
+    return zero(u_inner)
+end
+
+@inline function penalty(u_outer, u_inner, inv_h, x, t,
+                         equations_parabolic::AbstractLaplaceDiffusion,
+                         dg::ParabolicFormulationLocalDG)
+    diffusivity = penalty_diffusivity(u_outer, u_inner, x, t,
+                                      equations_parabolic)
+    return inv_h .* dg.penalty_parameter .* (u_outer - u_inner) .* diffusivity
+end
+
 include("laplace_diffusion_1d.jl")
 include("laplace_diffusion_2d.jl")
 include("laplace_diffusion_3d.jl")